Disable aws-lc's codegen.
Background: aws-lc uses codegen to create some of its source files, and it runs these generators if the user's machine has Go and Perl installed. However, up-to-date versions of the generated files are already committed to the repo, so it's not necessary for users to run codegen on their machines.

This codegen has led to unexpected issues (ex: failure due to network access at build time) so let's just turn it off and reduce unnecessary complexity.