CI job for OpenBSD (#134)
Add a CI job for OpenBSD. This is possible thanks to support being added for OpenBSD to crt-builder. This requires using a custom action to get an OpenBSD instance running (a VM hosted on macOS). Unfortunately, the custom action doesn't do all the necessary things to setup the environment for us so the job has to do some prep work before initiating the build.

---------

Co-authored-by: Michael Graeb <graebm@amazon.com>